move runtime executor plugin file to runtime executor, move docker runtime executor plugin def to own plugins file
